Update 20220114
Complete re-working of how data is read in from an XML file to the program. Code was made more compact by using built-in functions of C++. Comments have been added to all .cpp files where it was felt that was needed. Speed tests have been performed on the old version compared to the new version, and there is a very minor improvement in performance. The biggest snag for increasing program performance currently is the current implementation of the Crypto++ hashing functions, but right now I would consider it acceptable.